#1780 extract duplicate code to a superclass CaseSensitiveTextCondition
