+++ src/main/java/com/codeborne/selenide/conditions/CaseSensitiveText.java
@@ -1,8 +1,6 @@
 package com.codeborne.selenide.conditions;
 
 import com.codeborne.selenide.Driver;
-import com.codeborne.selenide.TextCheck;
-import com.codeborne.selenide.impl.Html;
 import org.openqa.selenium.WebElement;
 
 import javax.annotation.CheckReturnValue;
@@ -10,23 +8,10 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 @ParametersAreNonnullByDefault
-public class CaseSensitiveText extends TextCondition {
+public class CaseSensitiveText extends CaseSensitiveTextCondition {
 
   public CaseSensitiveText(String expectedText) {
     super("text case sensitive", expectedText);
-  }
-
-  @Override
-  protected boolean match(String actualText, String expectedText) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  protected boolean match(TextCheck textCheck, String actualText, String expectedText) {
-    return switch (textCheck) {
-      case FULL_TEXT -> Html.text.equalsCaseSensitive(actualText, expectedText);
-      case PARTIAL_TEXT -> Html.text.containsCaseSensitive(actualText, expectedText);
-    };
   }
 
   @Nullable
