+++ src/main/java/com/codeborne/selenide/conditions/SelectedText.java
@@ -1,8 +1,6 @@
 package com.codeborne.selenide.conditions;
 
 import com.codeborne.selenide.Driver;
-import com.codeborne.selenide.TextCheck;
-import com.codeborne.selenide.impl.Html;
 import com.codeborne.selenide.impl.JavaScript;
 import org.openqa.selenium.WebElement;
 
@@ -11,25 +9,11 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 @ParametersAreNonnullByDefault
-public class SelectedText extends TextCondition {
+public class SelectedText extends CaseSensitiveTextCondition {
   private final JavaScript js = new JavaScript("get-selected-text.js");
 
   public SelectedText(String expectedText) {
     super("selected text", expectedText);
-  }
-
-  @CheckReturnValue
-  @Override
-  protected boolean match(String actualText, String expectedText) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  protected boolean match(TextCheck textCheck, String actualText, String expectedText) {
-    return switch (textCheck) {
-      case FULL_TEXT -> Html.text.equalsCaseSensitive(actualText, expectedText);
-      case PARTIAL_TEXT -> Html.text.containsCaseSensitive(actualText, expectedText);
-    };
   }
 
   @Nullable
